[Refactor] Restructure packages into global and domain layers
