+++ src/main/java/calculator/domain/calculator/service/CalculatorService.java
@@ -1,4 +1,4 @@
-package calculator.service;
+package calculator.domain.calculator.service;
 
 public interface CalculatorService {
     int calculate(String input);
